Added the bomb throwing ability
Added the ability to throw bombs by clicking the right mouse button

Throwing a bomb will create an explosion, which currently has a placeholder of a red circle.

Most of the bomb settings can be adjusted within the GrenadeObject prefab under HattPrefabs.

Added children with triggered 2d colliders to the floor and ceiling (and prevFloor and prevCeiling) so that thrown bombs would interact with them.

Changed the Jetpack Trail's Box Collider to be a trigger (Is Trigger set to true), and made the trees not a trigger (Is Trigger set to false), to allow for the bombs and the explosions the bombs create to interact properly with both the trees and the rest of the environment.

I don't entirely understand why the numbers behind throwing the bombs are working the way they do, but I'm mostly happy with it so have decided not to interrogate it further for the time being.